Move client audio files to thier own sub-directory
