+++ client/control.cpp
@@ -1,5 +1,5 @@
 /*__            ___                 ***************************************
-/   \          /   \          Copyright (c) 1996-2020 Freeciv21 and Freeciv
+/   \          /   \          Copyright (c) 1996-2023 Freeciv21 and Freeciv
 \_   \        /  __/          contributors. This file is part of Freeciv21.
  _\   \      /  /__     Freeciv21 is free software: you can redistribute it
  \___  \____/   __/    and/or modify it under the terms of the GNU  General
@@ -36,7 +36,7 @@
 #include "menu_g.h"
 
 // client
-#include "audio.h"
+#include "audio/audio.h"
 #include "client_main.h"
 #include "climap.h"
 #include "climisc.h"
