+++ client/dialogs.cpp
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 1996-2022 Freeciv21 and Freeciv contributors. This file is
+ Copyright (c) 1996-2023 Freeciv21 and Freeciv contributors. This file is
  part of Freeciv21. Freeciv21 is free software: you can redistribute it
  and/or modify it under the terms of the GNU  General Public License  as
  published by the Free Software Foundation, either version 3 of the
@@ -34,7 +34,7 @@
 #include "research.h"
 #include "style.h"
 // client
-#include "audio.h"
+#include "audio/audio.h"
 #include "chatline_common.h"
 #include "client_main.h"
 #include "control.h"
