+++ client/packhand.cpp
@@ -1,5 +1,5 @@
 /*__            ___                 ***************************************
-/   \          /   \          Copyright (c) 1996-2020 Freeciv21 and Freeciv
+/   \          /   \          Copyright (c) 1996-2023 Freeciv21 and Freeciv
 \_   \        /  __/          contributors. This file is part of Freeciv21.
  _\   \      /  /__     Freeciv21 is free software: you can redistribute it
  \___  \____/   __/    and/or modify it under the terms of the GNU  General
@@ -77,7 +77,7 @@
 
 // client
 #include "attribute.h"
-#include "audio.h"
+#include "audio/audio.h"
 #include "chatline_common.h"
 #include "citydlg.h"
 #include "citydlg_common.h"
